IRS Plan page as a type of Intervention Plan page
