+++ src/constants.tsx
@@ -83,8 +83,8 @@ export const LOGOUT_URL = '/logout';
 export type LOGOUT_URL = typeof LOGOUT_URL;
 export const HOME_URL = '/';
 export type HOME_URL = typeof HOME_URL;
-export const IRS_URL = '/irs';
-export type IRS_URL = typeof IRS_URL;
+export const INTERVENTION_IRS_URL = '/intervention/irs';
+export type INTERVENTION_IRS_URL = typeof INTERVENTION_IRS_URL;
 export const FI_URL = '/focus-investigation';
 export type FI_URL = typeof FI_URL;
 export const FI_SINGLE_URL = '/focus-investigation/view';
