+++ src/containers/pages/InterventionPlan/IRS/index.tsx
@@ -1,0 +1,18 @@
+// this is the IRS page component
+import * as React from 'react';
+
+class IRS extends React.Component<{}, {}> {
+  constructor(props: {}) {
+    super(props);
+  }
+
+  public render() {
+    return (
+      <div>
+        <h1>IRS</h1>
+      </div>
+    );
+  }
+}
+
+export default IRS;
